+++ extensions/chroma/environment/converter.ts
@@ -4,6 +4,8 @@ import { Vector3 } from '../../../types/beatmap/shared/heck.ts';
 
 export function envV2toV3(env: IChromaEnvironmentV2[]): IChromaEnvironment[] {
     return env.map((e) => {
+        let components = {};
+        if (e._lightID) components = { ILightWithId: { lightID: e._lightID } };
         if (e._id && e._lookupMethod) {
             return {
                 id: e._id,
@@ -16,19 +18,19 @@ export function envV2toV3(env: IChromaEnvironmentV2[]): IChromaEnvironment[] {
                 rotation: e._rotation,
                 localPosition: e._localPosition?.map((n) => n * 0.6) as Vector3,
                 localRotation: e._localRotation,
-                lightID: e._lightID,
-            };
+                components,
+            } as IChromaEnvironment;
         }
         if (e._geometry) {
             return {
                 geometry: e._geometry.map((g) => {
                     return {
                         type: g._type,
+                        material: g._material,
                         spawnCount: g._spawnCount,
                         track: g._track,
-                        shaderPreset: g._shader,
-                        shaderKeywords: g._shaderKeywords,
                         collision: g._collision,
+                        color: g._color,
                     };
                 }),
                 track: e._track,
@@ -39,8 +41,8 @@ export function envV2toV3(env: IChromaEnvironmentV2[]): IChromaEnvironment[] {
                 rotation: e._rotation,
                 localPosition: e._localPosition?.map((n) => n * 0.6) as Vector3,
                 localRotation: e._localRotation,
-                lightID: e._lightID,
-            };
+                components,
+            } as IChromaEnvironment;
         }
         throw new Error('Error converting environment v2 to v3');
     });
@@ -60,19 +62,19 @@ export function envV3toV2(env: IChromaEnvironment[]): IChromaEnvironmentV2[] {
                 _rotation: e.rotation,
                 _localPosition: e.localPosition?.map((n) => n / 0.6) as Vector3,
                 _localRotation: e.localRotation,
-                _lightID: e.lightID,
-            };
+                _lightID: e.components?.ILightWithId?.lightID,
+            } as IChromaEnvironmentV2;
         }
         if (e.geometry) {
             return {
                 _geometry: e.geometry.map((g) => {
                     return {
                         _type: g.type,
+                        _material: g.material,
                         _spawnCount: g.spawnCount,
                         _track: g.track,
-                        _shaderPreset: g.shader,
-                        _shaderKeywords: g.shaderKeywords,
                         _collision: g.collision,
+                        _color: g.color,
                     };
                 }),
                 _track: e.track,
@@ -83,8 +85,8 @@ export function envV3toV2(env: IChromaEnvironment[]): IChromaEnvironmentV2[] {
                 _rotation: e.rotation,
                 _localPosition: e.localPosition?.map((n) => n / 0.6) as Vector3,
                 _localRotation: e.localRotation,
-                _lightID: e.lightID,
-            };
+                _lightID: e.components?.ILightWithId?.lightID,
+            } as IChromaEnvironmentV2;
         }
         throw new Error('Error converting environment v3 to v2');
     });